requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
aiohttp==3.9.3
numpy==1.26.4
pydantic==2.6.1
beautifulsoup4
html2text
tqdm
markdown
python-multipart
setuptools
